requirements.txt
# Python
requests

# Django
Django
django-filter
djangorestframework
drf-yasg
